chore: refactor gen subcommands
Move them to own package.
