+++ cmd/talosctl/cmd/mgmt/gen/config.go
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-package mgmt
+package gen
 
 import (
 	"fmt"
@@ -257,7 +257,7 @@ func writeV1Alpha1Config(args []string) error {
 }
 
 func init() {
-	genCmd.AddCommand(genConfigCmd)
+	Cmd.AddCommand(genConfigCmd)
 	genConfigCmd.Flags().StringVar(&genConfigCmdFlags.installDisk, "install-disk", "/dev/sda", "the disk to install to")
 	genConfigCmd.Flags().StringVar(&genConfigCmdFlags.installImage, "install-image", helpers.DefaultImage(images.DefaultInstallerImageRepository), "the image used to perform an installation")
 	genConfigCmd.Flags().StringSliceVar(&genConfigCmdFlags.additionalSANs, "additional-sans", []string{}, "additional Subject-Alt-Names for the APIServer certificate")
